Add files via upload
Analyzing crime data from 2020 to the present is essential for understanding trends and patterns in criminal activity. Such datasets are valuable for law enforcement agencies, policymakers, researchers, and the public to make informed decisions about crime prevention and intervention strategies. When working with a crime dataset spanning this time frame, it's important to consider several factors to determine whether it's suitable for crime analysis